modified he code so that the error pone compiler fails the build. Made some changes to the pom file

+++ src/test/java/org/zrgs/errorprone/ShortSetTest.java
@@ -28,12 +28,11 @@ public class ShortSetTest
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
   @Test
-  public void testShortSet()
-  {
+  public void testShortSet() {
 
     Set<Short> s = new HashSet<>();
 
-    for (short i = 0; i < 100; i++)
+    for (int i = 0; i < 100; i++)
     {
       s.add(i);
       s.remove(i - 1);
